everytime/requirements.txt
asgiref==3.4.1
Django==3.2.6
mysqlclient==2.0.1
pytz==2021.1
sqlparse==0.4.1
django-filter==21.1
django-debug-toolbar==2.2
djangorestframework==3.11.1
djangorestframework-simplejwt==5.0.0
django-rest-authtoken
Pillow  # profile picture 관련, 이미지처리
drf-yasg
django-storages # 다양한 storage 활용을 위한
boto3 # aws s3 접근을 위
faker
django-cors-headers
django-multiselectfield
django-extensions
ipython[notebook]
numpy==1.22.0
pandas==1.3.5
xlrd==2.0.1